pyproject.toml
[tool.poetry]
name = "sim"
version = "0.1.0"
description = ""
authors = ["nohvenell2 <nohvenell2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
sqlalchemy = "^2.0.37"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
